# 99_MyFiles/Project_Trial/Challenge_5_v2/contracts/api-contract.yaml
openapi: 3.1.0
info:
  title: SharkNinja Consumer Sentiment Dashboard
  description: Consumer-electronics-aware NLP pipeline for social listening. Handles
    sarcasm, comparative sentiment, ABSA, and CE domain terminology.
  version: 1.0.0
paths:
  /api/mentions:
    get:
      tags:
      - mentions
      summary: List Mentions
      description: List mentions with optional filtering. Supports drill-through from
        all dashboard pages.
      operationId: list_mentions_api_mentions_get
      parameters:
      - name: platform
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/SourcePlatform'
          - type: 'null'
          title: Platform
      - name: brand
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Brand'
          - type: 'null'
          title: Brand
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Category'
          - type: 'null'
          title: Category
      - name: product_model
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Model
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Date To
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: topic_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Topic Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentionListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/mentions/{mention_id}:
    get:
      tags:
      - mentions
      summary: Get Mention
      description: Retrieve a single mention by ID.
      operationId: get_mention_api_mentions__mention_id__get
      parameters:
      - name: mention_id
        in: path
        required: true
        schema:
          type: string
          title: Mention Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mention'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/overview/kpis:
    get:
      tags:
      - overview
      summary: Get Kpis
      description: 'Top-line KPIs: total mentions, overall sentiment score, WoW delta,
        rising aspects.'
      operationId: get_kpis_api_overview_kpis_get
      parameters:
      - name: brand
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Brand'
          - type: 'null'
          title: Brand
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Category'
          - type: 'null'
          title: Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewKPIs'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/overview/timeseries:
    get:
      tags:
      - overview
      summary: Get Timeseries
      description: Daily sentiment timeseries, filterable by brand/category.
      operationId: get_timeseries_api_overview_timeseries_get
      parameters:
      - name: brand
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Brand'
          - type: 'null'
          title: Brand
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Category'
          - type: 'null'
          title: Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeseriesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/overview/share_of_voice:
    get:
      tags:
      - overview
      summary: Get Share Of Voice
      description: Share-of-voice by brand. Optionally filtered by category.
      operationId: get_share_of_voice_api_overview_share_of_voice_get
      parameters:
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Category'
          - type: 'null'
          title: Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareOfVoiceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/alerts:
    get:
      tags:
      - alerts
      summary: List Alerts
      description: 'Rising negative aspect alerts sorted by severity.

        Use ?acknowledged=false for active alerts, ?acknowledged=true for history.

        Use ?platforms=reddit&platforms=amazon to filter by source platform.'
      operationId: list_alerts_api_alerts_get
      parameters:
      - name: brand
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Brand'
          - type: 'null'
          title: Brand
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Category'
          - type: 'null'
          title: Category
      - name: acknowledged
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Filter by acknowledged state. ?acknowledged=false for active,
            ?acknowledged=true for history
          title: Acknowledged
        description: Filter by acknowledged state. ?acknowledged=false for active,
          ?acknowledged=true for history
      - name: platforms
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              $ref: '#/components/schemas/SourcePlatform'
          - type: 'null'
          description: Filter alerts to those involving any of the specified platforms
          title: Platforms
        description: Filter alerts to those involving any of the specified platforms
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/alerts/{alert_id}/acknowledge:
    patch:
      tags:
      - alerts
      summary: Acknowledge Alert
      description: Mark an alert as acknowledged.
      operationId: acknowledge_alert_api_alerts__alert_id__acknowledge_patch
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertAcknowledgeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AlertAcknowledgeResponse:
      properties:
        alert_id:
          type: string
          title: Alert Id
        acknowledged:
          type: boolean
          title: Acknowledged
      type: object
      required:
      - alert_id
      - acknowledged
      title: AlertAcknowledgeResponse
    AlertEvent:
      properties:
        alert_id:
          type: string
          title: Alert Id
        product_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Model
        brand:
          $ref: '#/components/schemas/Brand'
        aspect:
          type: string
          title: Aspect
        severity:
          type: number
          title: Severity
          description: Composite severity score
        score_drop:
          type: number
          title: Score Drop
          description: How much the aspect score dropped
        triggered_at:
          type: string
          format: date-time
          title: Triggered At
        acknowledged:
          type: boolean
          title: Acknowledged
          default: false
        exemplar_mentions:
          items:
            $ref: '#/components/schemas/Mention'
          type: array
          title: Exemplar Mentions
        platforms:
          items:
            $ref: '#/components/schemas/SourcePlatform'
          type: array
          title: Platforms
          description: Distinct set of source platforms from exemplar_mentions[*].source_platform
      type: object
      required:
      - alert_id
      - brand
      - aspect
      - severity
      - score_drop
      - triggered_at
      title: AlertEvent
    AlertListResponse:
      properties:
        total:
          type: integer
          title: Total
        items:
          items:
            $ref: '#/components/schemas/AlertEvent'
          type: array
          title: Items
      type: object
      required:
      - total
      - items
      title: AlertListResponse
    AspectSentiment:
      properties:
        name:
          type: string
          title: Name
          description: Canonical aspect name, e.g. 'suction', 'battery', 'brushroll'
        polarity:
          $ref: '#/components/schemas/Polarity'
        score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Score
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
        snippet:
          type: string
          title: Snippet
          description: Text fragment that triggered this aspect extraction
      type: object
      required:
      - name
      - polarity
      - score
      - confidence
      - snippet
      title: AspectSentiment
...
    Brand:
      type: string
      enum:
      - shark
      - ninja
      - dyson
      - irobot
      - roborock
      - kitchenaid
      - breville
      - cuisinart
      - keurig
      - delonghi
      - other
      title: Brand
    BrandShare:
      properties:
        brand:
          $ref: '#/components/schemas/Brand'
        mention_count:
          type: integer
          title: Mention Count
        share:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Share
          description: Fraction of total mentions
      type: object
      required:
      - brand
      - mention_count
      - share
      title: BrandShare
    Category:
      type: string
      enum:
      - robot_vacuum
      - cordless_stick
      - upright
      - air_fryer
      - pressure_cooker
      - blender
      - ice_cream_maker
      - coffee
      - air_purifier
      - hair_tool
      - other
      title: Category
    ComparativePair:
      properties:
        brand:
          $ref: '#/components/schemas/Brand'
        aspect:
          type: string
          title: Aspect
          description: Canonical aspect name
        polarity:
          $ref: '#/components/schemas/Polarity'
        score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Score
      type: object
      required:
      - brand
      - aspect
      - polarity
      - score
      title: ComparativePair
...
    DerivedSentiment:
      properties:
        overall_sentiment:
          $ref: '#/components/schemas/SentimentLabel'
        compound_score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Compound Score
          description: Single scalar for charting
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          description: Pipeline confidence in this result
        sarcasm_flag:
          type: boolean
          title: Sarcasm Flag
          description: True if sarcasm heuristics fired
        aspects:
          items:
            $ref: '#/components/schemas/AspectSentiment'
          type: array
          title: Aspects
        comparative_pairs:
          anyOf:
          - items:
              $ref: '#/components/schemas/ComparativePair'
            type: array
          - type: 'null'
          title: Comparative Pairs
          description: Only populated when a comparative claim is detected
      type: object
      required:
      - overall_sentiment
      - compound_score
      - confidence
      - sarcasm_flag
      title: DerivedSentiment
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Mention:
      properties:
        mention_id:
          type: string
          title: Mention Id
          description: Stable UUID identifier
        source_platform:
          $ref: '#/components/schemas/SourcePlatform'
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
          description: URL to original mention; null for closed sources
        author_handle:
          anyOf:
          - type: string
          - type: 'null'
          title: Author Handle
          description: Anonymized/hashed author handle
        posted_at:
          type: string
          format: date-time
          title: Posted At
          description: When the mention was published (UTC)
        ingested_at:
          type: string
          format: date-time
          title: Ingested At
          description: When our pipeline processed it (UTC)
        brand:
          $ref: '#/components/schemas/Brand'
        category:
          $ref: '#/components/schemas/Category'
        product_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Model
          description: Specific SKU name if available
        text:
          type: string
          title: Text
          description: Raw mention text, pre-NLP
        rating:
          anyOf:
          - type: number
          - type: 'null'
          title: Rating
          description: Source-provided star rating if available
        language:
          type: string
          title: Language
          description: ISO-639-1 language code
          default: en
        derived:
          $ref: '#/components/schemas/DerivedSentiment'
        record_type:
          type: string
          enum:
          - review
          - recall
          title: Record Type
          description: '''review'' for normal mentions; ''recall'' for CPSC recall
            records'
          default: review
      type: object
      required:
      - mention_id
      - source_platform
      - posted_at
      - ingested_at
      - brand
      - category
      - text
      - derived
      title: Mention
...
    MentionListResponse:
      properties:
        total:
          type: integer
          title: Total
        items:
          items:
            $ref: '#/components/schemas/Mention'
          type: array
          title: Items
      type: object
      required:
      - total
      - items
      title: MentionListResponse
    OverviewKPIs:
      properties:
        total_mentions:
          type: integer
          title: Total Mentions
        overall_score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Overall Score
        wow_delta:
          type: number
          title: Wow Delta
          description: Week-over-week score change
        top_rising_negative:
          items:
            type: string
          type: array
          title: Top Rising Negative
          description: Aspect names with worst score drops
        top_rising_positive:
          items:
            type: string
          type: array
          title: Top Rising Positive
          description: Aspect names with best score gains
      type: object
      required:
      - total_mentions
      - overall_score
      - wow_delta
      - top_rising_negative
      - top_rising_positive
      title: OverviewKPIs
...
    Polarity:
      type: string
      enum:
      - positive
      - negative
      - neutral
      - mixed
      title: Polarity
...
    SentimentLabel:
      type: string
      enum:
      - positive
      - negative
      - neutral
      - mixed
      title: SentimentLabel
...
    ShareOfVoiceResponse:
      properties:
        total_mentions:
          type: integer
          title: Total Mentions
        brands:
          items:
            $ref: '#/components/schemas/BrandShare'
          type: array
          title: Brands
      type: object
      required:
      - total_mentions
      - brands
      title: ShareOfVoiceResponse
...
    SourcePlatform:
      type: string
      enum:
      - reddit
      - amazon
      - youtube
      - trustpilot
      - twitter
      - other
      title: SourcePlatform
    TimeseriesPoint:
      properties:
        date:
          type: string
          title: Date
          description: ISO date YYYY-MM-DD
        score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Score
        mention_count:
          type: integer
          title: Mention Count
      type: object
      required:
      - date
      - score
      - mention_count
      title: TimeseriesPoint
    TimeseriesResponse:
      properties:
        series:
          items:
            $ref: '#/components/schemas/TimeseriesPoint'
          type: array
          title: Series
        brand:
          anyOf:
          - $ref: '#/components/schemas/Brand'
          - type: 'null'
        category:
          anyOf:
          - $ref: '#/components/schemas/Category'
          - type: 'null'
      type: object
      required:
      - series
      title: TimeseriesResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError